move log_entry_types registry to separate file
