+++ src/pretix/base/logentrytypes.py
@@ -25,131 +25,10 @@ from django.urls import reverse
 from django.utils.html import escape
 from django.utils.translation import gettext_lazy as _, pgettext_lazy
 
-from pretix.base.signals import EventPluginRegistry
-
-
-def make_link(a_map, wrapper, is_active=True, event=None, plugin_name=None):
-    if a_map:
-        if is_active:
-            a_map['val'] = '<a href="{href}">{val}</a>'.format_map(a_map)
-        elif event and plugin_name:
-            a_map['val'] = (
-                '<i>{val}</i> <a href="{plugin_href}">'
-                '<span data-toggle="tooltip" title="{errmes}" class="fa fa-warning fa-fw"></span></a>'
-            ).format_map({
-                **a_map,
-                "errmes": _("The relevant plugin is currently not active. To activate it, click here to go to the plugin settings."),
-                "plugin_href": reverse('control:event.settings.plugins', kwargs={
-                    'organizer': event.organizer.slug,
-                    'event': event.slug,
-                }) + '#plugin_' + plugin_name,
-            })
-        else:
-            a_map['val'] = '<i>{val}</i> <span data-toggle="tooltip" title="{errmes}" class="fa fa-warning fa-fw"></span>'.format_map({
-                **a_map,
-                "errmes": _("The relevant plugin is currently not active."),
-            })
-        return wrapper.format_map(a_map)
-
-
-class LogEntryTypeRegistry(EventPluginRegistry):
-    def __init__(self):
-        super().__init__({'action_type': lambda o: getattr(o, 'action_type')})
-
-    def register(self, *objs):
-        for obj in objs:
-            if not isinstance(obj, LogEntryType):
-                raise TypeError('Entries must be derived from LogEntryType')
-
-            if obj.__module__ == LogEntryType.__module__:
-                raise TypeError('Must not register base classes, only derived ones')
-
-        return super().register(*objs)
-
-    def new_from_dict(self, data):
-        """
-        Register multiple instance of a `LogEntryType` class with different `action_type`
-        and plain text strings, as given by the items of the specified data dictionary.
-
-        This method is designed to be used as a decorator as follows:
-
-        .. code-block:: python
-
-            @log_entry_types.new_from_dict({
-                'pretix.event.item.added': _('The product has been created.'),
-                'pretix.event.item.changed': _('The product has been changed.'),
-                # ...
-            })
-            class CoreItemLogEntryType(ItemLogEntryType):
-                # ...
-
-        :param data: action types and descriptions
-                     ``{"some_action_type": "Plain text description", ...}``
-        """
-        def reg(clz):
-            for action_type, plain in data.items():
-                self.register(clz(action_type=action_type, plain=plain))
-            return clz
-        return reg
-
-
-"""
-Registry for LogEntry types.
-
-Each entry in this registry should be an instance of a subclass of ``LogEntryType``.
-They are annotated with their ``action_type`` and the defining ``plugin``.
-"""
-log_entry_types = LogEntryTypeRegistry()
-
-
-class LogEntryType:
-    """
-    Base class for a type of LogEntry, identified by its action_type.
-    """
-
-    def __init__(self, action_type=None, plain=None):
-        if action_type:
-            self.action_type = action_type
-        if plain:
-            self.plain = plain
-
-    def display(self, logentry):
-        """
-        Returns the message to be displayed for a given logentry of this type.
-
-        :return: `str` or `LazyI18nString`
-        """
-        if hasattr(self, 'plain'):
-            plain = str(self.plain)
-            if '{' in plain:
-                data = defaultdict(lambda: '?', logentry.parsed_data)
-                return plain.format_map(data)
-            else:
-                return plain
-
-    def get_object_link_info(self, logentry) -> dict:
-        """
-        Return information to generate a link to the `content_object` of a given log entry.
-
-        Not implemented in the base class, causing the object link to be omitted.
-
-        :return: Dictionary with the keys ``href`` (containing a URL to view/edit the object) and ``val`` (containing the
-                 escaped text for the anchor element)
-        """
-        pass
-
-    def get_object_link(self, logentry):
-        a_map = self.get_object_link_info(logentry)
-        return make_link(a_map, self.object_link_wrapper)
-
-    object_link_wrapper = '{val}'
-
-    def shred_pii(self, logentry):
-        """
-        To be used for shredding personally identified information contained in the data field of a LogEntry of this
-        type.
-        """
-        raise NotImplementedError
+from .logentrytype_registry import (  # noqa
+    ClearDataShredderMixin, LogEntryType, NoOpShredderMixin, log_entry_types,
+    make_link, LogEntryTypeRegistry,
+)
 
 
 class EventLogEntryType(LogEntryType):
@@ -241,13 +120,3 @@ class TaxRuleLogEntryType(EventLogEntryType):
     object_link_wrapper = _('Tax rule {val}')
     object_link_viewname = 'control:event.settings.tax.edit'
     object_link_argname = 'rule'
-
-
-class NoOpShredderMixin:
-    def shred_pii(self, logentry):
-        pass
-
-
-class ClearDataShredderMixin:
-    def shred_pii(self, logentry):
-        logentry.data = None
